Get adjusted means with compact letter display

This function is a wrapper for emmeans::emmeans() and multcomp::cld() (details here). Based on a given linear model object, it computes estimated marginal means (a.k.a. least-squares means, adjusted means, modelbased means) for specified factors or factor combinations as well as comparisons or contrasts among them. Moreover, it adds the compact letter display to the comparisons/contrasts. Finally, its output is a list with six elements: for the means and their diffs, respectively, (i) a tibble with the results, (ii) the same tibble but in a print-ready format and (iii) an info text summarizing the results.

emmeans_BM(
  model,
  specs_string,
  lmer.df = c("Asymptotic", "Satterthwaite", "Kenward-Roger")[3],
  adjust = c("none", "tukey", "bonferroni")[1],
  out_lang = c("ger", "eng")[1],
  num_accuracy = 0.01,
  sort = c("none", "asc", "desc")[1]
)

Arguments

model

a linear model object. See the list of supported model classes

specs_string

a character vector specifying the names of the predictors over which emmeans are desired

lmer.df

degrees of freedom method. See the list of available methods

adjust

multiplicity adjustment. Follow this link and scroll down o to the “P-value adjustments” heading within the “summary.emmGrid” section for more details on e.g. Fisher’s LSD test, Tukey-test, Bonferroni adjustment etc.

out_lang

language for the formatted result table and text

num_accuracy

format the number of digits shown after the decimal separator

sort

sort the levels of the predictors given in specs_string in the output

Details

Here is a discussion on the speed of calculating confidence intervals with different degrees of freedom methods.

Examples

library(BioMathR)
pigsmod <- lm(conc ~ source, data = emmeans::pigs)

emmeans_BM(model = pigsmod, specs_string = "~ source", out_lang = "eng")
#> $means
#> # A tibble: 3 × 11
#>   source emmean   SEM    df lower…¹ upper…² .group emmea…³ SEM_fmt df_fmt CI_fmt
#>   <fct>   <dbl> <dbl> <dbl>   <dbl>   <dbl> <chr>  <chr>   <chr>   <chr>  <glue>
#> 1 fish     30.4  2.02    26    26.2    34.5 a      30.36   2.02    26     [26.2…
#> 2 soy      38.3  2.02    26    34.2    42.5 b      38.34   2.02    26     [34.1…
#> 3 skim     44.1  2.13    26    39.8    48.5 b      44.14   2.13    26     [39.7…
#> # … with abbreviated variable names ¹​lower.CL, ²​upper.CL, ³​emmean_fmt
#> 
#> $diffs
#> # A tibble: 3 × 12
#>   contrast    estim…¹   SED    df lower…² upper…³ p.value estim…⁴ SED_fmt df_fmt
#>   <chr>         <dbl> <dbl> <dbl>   <dbl>   <dbl>   <dbl> <chr>   <chr>   <chr> 
#> 1 fish - soy    -7.98  2.85    26   -13.8  -2.11  9.61e-3 -7.98   2.85    26    
#> 2 fish - skim  -13.8   2.93    26   -19.8  -7.76  7.42e-5 -13.78  2.93    26    
#> 3 soy - skim    -5.80  2.93    26   -11.8   0.224 5.85e-2 -5.80   2.93    26    
#> # … with 2 more variables: CI_fmt <glue>, p.value_fmt <chr>, and abbreviated
#> #   variable names ¹​estimate, ²​lower.CL, ³​upper.CL, ⁴​estimate_fmt
#> 
#> $means_print
#> # A tibble: 3 × 6
#>   source Mean  SEM   DF    `95% Conf. Int.` CLD  
#>   <fct>  <chr> <chr> <chr> <glue>           <chr>
#> 1 fish   30.36 2.02  26    [26.21; 34.51]   a    
#> 2 soy    38.34 2.02  26    [34.19; 42.49]   b    
#> 3 skim   44.14 2.13  26    [39.77; 48.52]   b    
#> 
#> $diffs_print
#> # A tibble: 3 × 6
#>   Contrast    Difference SED   DF    `95% Conf. Int.` `p-value`  
#>   <chr>       <chr>      <chr> <chr> <glue>           <chr>      
#> 1 fish - soy  -7.98      2.85  26    [-13.85; -2.11]  p = 0.010**
#> 2 fish - skim -13.78     2.93  26    [-19.81; -7.76]  p < .001***
#> 3 soy - skim  -5.80      2.93  26    [-11.83; 0.22]   p = 0.058  
#> 
#> $means_info
#> [1] "Modelbased means for source (df according to Kenward-Roger method). Regarding the CLD, it is true that means not sharing any letter are significantly different according to the Fisher's LSD Test at the 5% level of significance. Abb.: Standard Error of the Mean (SEM); Degrees of Freedom (DF); Compact Letter Display (CLD)"
#> 
#> $diffs_info
#> [1] "Modelbased mean comparisons for source (df according to Kenward-Roger method). Contrasts with a p-value < 0.05 are statistically significant accordint to the Fisher's LSD Test at the 5% level of significance, i.e. the difference between the respective means is significantly different from 0. Abb.: Standard Error of the Mean (SEM); Degrees of Freedom (DF); Compact Letter Display (CLD)"
#> 

emmeans_BM(model = pigsmod, specs_string = "~ source", out_lang = "ger")
#> $means
#> # A tibble: 3 × 11
#>   source emmean   SEM    df lower…¹ upper…² .group emmea…³ SEM_fmt df_fmt CI_fmt
#>   <fct>   <dbl> <dbl> <dbl>   <dbl>   <dbl> <chr>  <chr>   <chr>   <chr>  <chr> 
#> 1 fish     30.4  2.02    26    26.2    34.5 a      30,36   2,02    26     [26,2…
#> 2 soy      38.3  2.02    26    34.2    42.5 b      38,34   2,02    26     [34,1…
#> 3 skim     44.1  2.13    26    39.8    48.5 b      44,14   2,13    26     [39,7…
#> # … with abbreviated variable names ¹​lower.CL, ²​upper.CL, ³​emmean_fmt
#> 
#> $diffs
#> # A tibble: 3 × 12
#>   contrast    estim…¹   SED    df lower…² upper…³ p.value estim…⁴ SED_fmt df_fmt
#>   <chr>         <dbl> <dbl> <dbl>   <dbl>   <dbl>   <dbl> <chr>   <chr>   <chr> 
#> 1 fish - soy    -7.98  2.85    26   -13.8  -2.11  9.61e-3 -7,98   2,85    26    
#> 2 fish - skim  -13.8   2.93    26   -19.8  -7.76  7.42e-5 -13,78  2,93    26    
#> 3 soy - skim    -5.80  2.93    26   -11.8   0.224 5.85e-2 -5,80   2,93    26    
#> # … with 2 more variables: CI_fmt <chr>, p.value_fmt <chr>, and abbreviated
#> #   variable names ¹​estimate, ²​lower.CL, ³​upper.CL, ⁴​estimate_fmt
#> 
#> $means_print
#> # A tibble: 3 × 6
#>   source Mittelwert SEM   FG    `95%-Vertrauensint.` CLD  
#>   <fct>  <chr>      <chr> <chr> <chr>                <chr>
#> 1 fish   30,36      2,02  26    [26,21; 34,51]       a    
#> 2 soy    38,34      2,02  26    [34,19; 42,49]       b    
#> 3 skim   44,14      2,13  26    [39,77; 48,52]       b    
#> 
#> $diffs_print
#> # A tibble: 3 × 6
#>   Kontrast    Differenz SED   FG    `95%-Vertrauensint.` `p-Wert`   
#>   <chr>       <chr>     <chr> <chr> <chr>                <chr>      
#> 1 fish - soy  -7,98     2,85  26    [-13,85; -2,11]      p = 0,010**
#> 2 fish - skim -13,78    2,93  26    [-19,81; -7,76]      p < ,001***
#> 3 soy - skim  -5,80     2,93  26    [-11,83; 0,22]       p = 0,058  
#> 
#> $means_info
#> [1] "Modellbasierte Mittelwerte: source (FG nach Kenward-Roger Methode). Es gilt dass Mittelwerte, die keinen gemeinsamen CLD-Buchstaben aufweisen, sich laut Fishers LSD Test statistisch signifikant voneinander unterscheiden (5% Signifikanzniveau). Abk.: Standard Error of the Mean (SEM); Freiheitsgrade (FG); Compact Letter Display (CLD)"
#> 
#> $diffs_info
#> [1] "Modellbasierte Mittelwertvergleiche: source (FG nach Kenward-Roger Methode). Kontraste mit einem p-Wert < 0.05 sind laut Fishers LSD Test statistisch signifikant, d.h. die Differenz zwischen den jeweiligen Mittelwerten ist statistisch signifikant verschieden von 0. Abk.: Standard Error of Difference (SED); Freiheitsgrade (FG); Compact Letter Display (CLD)"
#>