Compute descriptive statistics

This function computes common descriptive statistics of numeric variables for grouped or ungrouped data and converts it to a tidy formatted table. This function tried to copy dlookr::describe() because of this issue.

describe(data, yvars, ungroupafter = TRUE, lang = c("eng", "ger")[1])

Arguments

data

a data.frame or a tbl_df or a grouped_df.

yvars

a character vector with the names of the columns to be described.

ungroupafter

If TRUE, the results will be ungroup()-ed.

lang

If "eng" (default) or "ger" columns will be renamed accordingly.

Value

A tibble

Examples

library(dplyr)
library(BioMathR)
PlantGrowth %>%
  group_by(group) %>%
  describe("weight")
#> # A tibble: 3 × 10
#>   Variable group     N  Miss  Mean StdDev   IQR   Min Median   Max
#>   <chr>    <fct> <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl>  <dbl> <dbl>
#> 1 weight   ctrl     10     0  5.03  0.583 0.743  4.17   5.15  6.11
#> 2 weight   trt1     10     0  4.66  0.794 0.662  3.59   4.55  6.03
#> 3 weight   trt2     10     0  5.53  0.443 0.467  4.92   5.44  6.31

PlantGrowth %>%
  mutate(weight2 = weight + 1) %>%
  group_by(group) %>%
  describe(c("weight", "weight2"))
#> # A tibble: 6 × 10
#>   Variable group     N  Miss  Mean StdDev   IQR   Min Median   Max
#>   <chr>    <fct> <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl>  <dbl> <dbl>
#> 1 weight   ctrl     10     0  5.03  0.583 0.743  4.17   5.15  6.11
#> 2 weight   trt1     10     0  4.66  0.794 0.662  3.59   4.55  6.03
#> 3 weight   trt2     10     0  5.53  0.443 0.467  4.92   5.44  6.31
#> 4 weight2  ctrl     10     0  6.03  0.583 0.743  5.17   6.15  7.11
#> 5 weight2  trt1     10     0  5.66  0.794 0.662  4.59   5.55  7.03
#> 6 weight2  trt2     10     0  6.53  0.443 0.467  5.92   6.44  7.31